% environment where values are stored

\name{env.gui.ea}

\alias{env.gui.ea}

\title{
  Internal variable
}

\description{
  Internal variable where the evolutionary algorithm object is stored
}

\usage{
  env.gui.ea
}


\value{
  The result of the function \code{new.env(FALSE,NULL)}.
}

\note{
}

\references{
}

\author{
  Kurt Sys (\email{kurt.sys@rug.ac.be})
}

\seealso{
  environment, new.env
}

\keyword{
  environment
}

\examples{
}
