% guis

\name{seperate guis}

\alias{genomestruc.gui}
\alias{newgen.gui}
\alias{nextgeneration.gui}
\alias{plotfitness.gui}
\alias{plotgene.gui}
\alias{save.ea.gui}

\title{
   Interactive windows for evolutionary algorithm functions.
}

\description{
  Set of functions which can be used as interface between user and
  evolutionary algorithm functions of package \code{seao}. 
}

\usage{
  genomestruc.gui()
  newgen.gui(ea=NULL)
  nextgeneration.gui(ea=NULL)
  plotfitness.gui(ea=NULL)
  plotgene.gui(ea=NULL)
  save.ea.gui(ea=NULL)
}

\arguments{
  \item{ea}{An object of type evoutionary algorithms. This
    contains the data to be plotted}
}

\details{
}


\references{
}

\author{
   Kurt Sys (\email{kurt.sys@rug.ac.be})
}

\seealso{
  \link[seao]{genomestruc}
  \link[seao]{newgen}
  \link[seao]{nextgen}
  \link[seao]{plotevolution.fit}
  \link[seao]{plotevolution.gene}
}

\keyword{
  GUI
}

\examples{
}
