% seao.guistart

\name{seao.guistart}

\alias{seao.guistart}

\title{
  Overall graphical user interface
}

\description{
  Gives a GUI for the evolutionary algorithms functions (of package \code{seao)}.
}

\usage{
  seao.guistart()
}

\note{
}

\references{
}

\author{
  Kurt Sys (\email{kurt.sys@rug.ac.be})
}

\seealso{
}

\keyword{
  GUI
}

\examples{
}
