% showvalues.gui

\name{showvalues.gui}

\alias{showvalues.gui}

\title{
  Internal function
}

\description{
  Small internal functions, which is actually the graphical interface
  (window) for showing fitnesses and alleles and editing the fitnesses
  of one generation.
}

\usage{
  showvalues.gui{ea=NULL, gen=NULL}
}

\arguments{
  \item{ea}{An object of type evoutionary algorithms. This
    contains the data to be plotted}
  \item{gen}{The generation to be shown and/or edited}
}

\details{
}


\references{
}

\author{
   Kurt Sys (\email{kurt.sys@rug.ac.be})
}

\seealso{
  \link{seao.guistart}
}

\keyword{
  GUI
}

\examples{
}
